webtester-core/src/test/java/info/novatec/testit/webtester/pageobjects/HeadlineTest.java
package info.novatec.testit.webtester.pageobjects;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

import org.junit.Test;
import org.mockito.InjectMocks;

import info.novatec.testit.webtester.AbstractPageObjectTest;


public class HeadlineTest extends AbstractPageObjectTest {

    @InjectMocks
    Headline cut;

    /* correctness of class */

    @Test
    public void testCorrectnessOfClassForWebElement_h1Tag() {
        stubWebElementTag("h1");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_h2Tag() {
        stubWebElementTag("h2");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_h3Tag() {
        stubWebElementTag("h3");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_h4Tag() {
        stubWebElementTag("h4");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_h5Tag() {
        stubWebElementTag("h5");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_h6Tag() {
        stubWebElementTag("h6");
        assertThatCorrectnessOfClassIs(true);
    }

    @Test
    public void testCorrectnessOfClassForWebElement_otherTag() {
        stubWebElementTag("other");
        assertThatCorrectnessOfClassIs(false);
    }

    /* utilities */

    private void assertThatCorrectnessOfClassIs(boolean expected) {
        assertThat(cut.isCorrectClassForWebElement(webElement), is(expected));
    }

}
